requirements.txt
Django==1.6.5
django-registration==1.0
django-twitter-feed==0.2
tweepy
south==1.0
django-taggit==0.12.2
django-taggit-templatetags==0.4.6dev
qhonuskan-votes==0.2
django-notify==1.1.2
django-mptt==0.6.1
sorl-thumbnail==11.12.1b
django-sekizai==0.7
wiki==0.0.24.2
django-haystack
whoosh==2.5.7